Bug workaround file managers.
Over the past few day she files-workspace script has beeb flaky.
This was a symptom of temperamental file manager apps misbehaving
indepedently of this project it seems,

At pre-release we need the minimum we can get awy witg. Thunar is
refusng to launch. PCManFM is losing filrd, PCManFM0-qt hasn't
succumbed to the plague.

RGe workaround is to dumb down the script to the minimum required using
pcmanfm-qt.