Update README.md with end-of-year ambition.
Tomorrow it will be one year since the fall that hospitalised me for eight weeks, causing the subsequent abandonment of the dopamine-2922 project. 

i am preparing a release of dopamine-2024 ready before 2024 and that means throwing everything over the fence. It's not best practice, but if I don't deliver there's nothing to show that you can get stuff done despite Parkinson's.