Menu key bottom row is Super_R
Although I've made a huge song-and-dance about the Menu key,
it has endedup on right backslash above Enter, while the
actual Menu key is lost in the 61 key arrow key shenanigans.
Anyway, it's surely the-right-thing to have separate left and
right modifier keys, hence the key with a Menu logo is now a
Super_R key for the right hand.