Visit workspace and launch a terminal.
This is a major enhancement to my way of working. THe ntrkeys table
has been taken over bt the system. The table is 500 to 599, it really
ought to be 300 to 399, but back compatibility.

I needed a way to say "put a terminak there" and this is it, it feels
natural. I'ver not been carried away, it's a minimal implementation, but
you needto be handy with a text editor to make it

Press Mod4+5 or Fn+ 5 or Enrter t o select the fifth table, then two
digits fothe workspace number.

i've snuck in a mix of xfce4-ermonal and qterminal